Cost calculation improvements

+++ src/main/java/com/sweetrpg/crafttracker/common/util/calc/ICostCalculator.java
@@ -2,6 +2,6 @@ package com.sweetrpg.crafttracker.common.util.calc;
 
 public interface ICostCalculator {
 
-    int calculate();
+    double calculate();
 
 }
